# WIP: Add caching vector to OpenSim::Model