This PR is for looking into caching the `Controller`s owned by a `Model` so that `computeControls` doesn't have to do an expensive `getComponentList` lookup. 

Basic profiling reveals that `getComponentList<T>` is somewhat expensive because it calls `dynamic_cast` on each element in the underlying component list. One way to get around this is to cache all `Controller`s pre-computation.