@adamkewley 2020-06-05
@tgeijten mentioned that, once `buildSystem` is called, a Model's component tree should be frozen, so I'll try adding the caching code just in there (rather than looking for `finalizeFromProperties` calls, which seem to dirty the cache) and I'll see if that works.