@adamkewley 2020-06-05
Errors printed with `CTEST_OUTPUT_ON_FAILURE=1 ctest`

```
==========================================================
 A Body without a Joint should get a Free (6dof) Joint    
==========================================================
[warning] Body 'block' not connected by a Joint.A FreeJoint will be added to connect it to ground.
Model::finalizeFromProperties() must be called first.
	In Object 'model' of type Model.
	Thrown at Model.cpp:1569 in printBasicInfo().
```

```
=============================================================
 OpenSim PinJoint vs. Simbody MobilizedBody::Pin 
=============================================================
Sockets for component knee of type [PinJoint] along with connectee paths:
    [PhysicalFrame]   child_frame : 
    [PhysicalFrame]  parent_frame : 
Inputs for component knee of type [PinJoint] along with connectee paths: none
Sockets for component knee of type [PinJoint] along with connectee paths:
    [PhysicalFrame]   child_frame : 
    [PhysicalFrame]  parent_frame : 
Inputs for component knee of type [PinJoint] along with connectee paths: none
hip equivalent FB_G = ~[~[0,0,838.15],~[0,0,0]]
knee equivalent FB_G = ~[~[602.036,362.353,329.975],~[0,0,0]]
****************************************************************************
uDot Error = 2.08124e-16: from body forces vs. mobility forces.
****************************************************************************
[info] Loaded model model from file testPinJointModel.osim
osim_state - sb_state: ~[0 0 0 0]

Simbody Final q's: ~[-0.497006 2.67136]
Simbody Final u's: ~[-2.13475 0.147729]

OpenSim Final q's: ~[-0.497006 2.67136]
OpenSim Final u's: ~[-2.13475 0.147729]

Simbody - OpenSim:  |q_err| = 8.9509e-16  |u_err| =2.50985e-15
```

```
=============================================================
 OpenSim SliderJoint vs. Simbody MobilizedBody::Slider 
=============================================================
testSliderJoint: testEquivalentBodyForceForGenForces
hip equivalent FB_G = ~[~[0,0,353.465],~[0,0,0]]
knee equivalent FB_G = ~[~[0,0,0],~[1.00378,-6.3934,147.328]]
****************************************************************************
uDot Error = 2.10769e-17: from body forces vs. mobility forces.
****************************************************************************
osim_state - sb_state: ~[0 0 0 0]

Simbody Final q's: ~[-0.827633 -0.108601]
Simbody Final u's: ~[-2.27122 0.566967]

OpenSim Final q's: ~[-0.827633 -0.108601]
OpenSim Final u's: ~[-2.27122 0.566967]

Simbody - OpenSim:  |q_err| = 0  |u_err| =0
```

```
=============================================================
 OpenSim PlanarJoint vs. Simbody MobilizedBody::Planar 
=============================================================
testPlanarJoint: testEquivalentBodyForceForGenForces
hip equivalent FB_G = ~[~[0,0,-214.933],~[-547.68,986.244,0]]
knee equivalent FB_G = ~[~[-43.0646,-40.0343,33.8498],~[208.562,-283.15,-69.5453]]
****************************************************************************
uDot Error = 2.72906e-16: from body forces vs. mobility forces.
****************************************************************************
osim_state - sb_state: ~[0 0 0 0 0 0 0 0 0 0 0 0]

Simbody Final q's: ~[-0.0790296 0.0436108 -1.04551 0.3405 0.44001 -0.331076]
Simbody Final u's: ~[0.367851 0.333212 -4.29282 -0.647212 0.707219 0.66492]

OpenSim Final q's: ~[-0.0790296 0.0436108 -1.04551 0.3405 0.44001 -0.331076]
OpenSim Final u's: ~[0.367851 0.333212 -4.29282 -0.647212 0.707219 0.66492]

Simbody - OpenSim:  |q_err| = 1.79208e-15  |u_err| =4.02523e-15
```

bla bla bla (same bs for loads of tests)


```
  testModelFinalizePropertiesAndConnections done. real/CPU ms: 21.8 / 21.8
Test failed due to exception: Component has no underlying System.
You must call initSystem() on the top-level Component (i.e. Model) first.
	In Object 'arm26' of type Model.
	Thrown at Component.cpp:1480 in getSystem().
```

```
[warning] Body 'Body1804289383' not connected by a Joint.A FreeJoint will be added to connect it to ground.
[warning] Body 'Body1303455736' not connected by a Joint.A FreeJoint will be added to connect it to ground.
*******************************************************
FAILURE: Body
testComponents:: model.initSystem() caused Model to no longer be up-to-date with its properties.
	Thrown at testComponents.cpp:330 in testComponent().
```

etc.

So, ignoring all the other noise, I'm guessing that `testComponents` is calling on `model` in such a way that the cache is over-caching and giving inconsistent results.