@adamkewley 2020-06-05
Alright, so it turns out that all of the `compute` methods are `const` (I didn't check this, and assumed they weren't), so I've been storing a `reference_wrapper` to a non-`const` `Controller`, which requires calling the non-`const` iterator (`updComponentList`), which returns the iterator *and* dirties some `_propertiesUpToDate` flag deep inside the `Object` base class, and that flag is used to dirty-check things, which is what makes those tests fail (they shouldn't be dirty).