@adamkewley 2020-06-05
Hm, seems to make a basically negligible performance difference?

Before:

```
          5,084.13 msec task-clock                #    0.996 CPUs utilized            ( +-  0.62% )
               514      context-switches          #  101.205 M/sec                    ( +-  6.69% )
                 0      cpu-migrations            #    0.025 M/sec                    ( +-100.00% )
            11,709      page-faults               # 2303.196 M/sec                    ( +-  0.01% )
    17,741,673,539      cycles                    # 3489879.231 GHz                   ( +-  0.64% )  (83.31%)
     1,006,683,885      stalled-cycles-frontend   #    5.67% frontend cycles idle     ( +-  2.06% )  (83.32%)
     1,611,382,140      stalled-cycles-backend    #    9.08% backend cycles idle      ( +-  0.58% )  (83.33%)
    26,317,797,673      instructions              #    1.48  insn per cycle         
                                                  #    0.06  stalled cycles per insn  ( +-  0.04% )  (83.33%)
     5,258,846,703      branches                  # 1034442429.899 M/sec              ( +-  0.03% )  (83.37%)
        72,321,154      branch-misses             #    1.38% of all branches          ( +-  3.30% )  (83.34%)

            5.1020 +- 0.0305 seconds time elapsed  ( +-  0.60% )
```

After:

```
 Performance counter stats for 'opensim-cmd run-tool setup_forward_tool.xml' (8 runs):

          5,083.10 msec task-clock                #    0.995 CPUs utilized            ( +-  0.40% )
               566      context-switches          #  111.387 M/sec                    ( +-  8.29% )
                 0      cpu-migrations            #    0.074 M/sec                    ( +- 70.15% )
            11,710      page-faults               # 2304.058 M/sec                    ( +-  0.01% )
    17,714,659,385      cycles                    # 3485422.407 GHz                   ( +-  0.28% )  (83.31%)
     1,010,644,205      stalled-cycles-frontend   #    5.71% frontend cycles idle     ( +-  1.09% )  (83.32%)
     1,615,578,421      stalled-cycles-backend    #    9.12% backend cycles idle      ( +-  0.90% )  (83.34%)
    26,305,026,323      instructions              #    1.48  insn per cycle         
                                                  #    0.06  stalled cycles per insn  ( +-  0.05% )  (83.34%)
     5,256,859,699      branches                  # 1034305892.573 M/sec              ( +-  0.02% )  (83.34%)
        70,432,593      branch-misses             #    1.34% of all branches          ( +-  0.97% )  (83.34%)

            5.1064 +- 0.0206 seconds time elapsed  ( +-  0.40% )
```