@adamkewley 2020-06-05
The model doesn't have many `Controllers`: the profile shows this. Switching to `ToyDropLanding`, I'm seeing:

before:

```text
 Performance counter stats for 'opensim-cmd run-tool setup_forward_tool.xml' (8 runs):

         15,772.55 msec task-clock                #    0.998 CPUs utilized            ( +-  0.14% )
             1,670      context-switches          #  105.859 M/sec                    ( +-  7.13% )
                 0      cpu-migrations            #    0.032 M/sec                    ( +- 75.59% )
             7,539      page-faults               #  477.979 M/sec                    ( +-  0.02% )
    54,478,009,764      cycles                    # 3454069.110 GHz                   ( +-  0.21% )  (83.32%)
     4,402,329,618      stalled-cycles-frontend   #    8.08% frontend cycles idle     ( +-  0.83% )  (83.33%)
     4,266,599,080      stalled-cycles-backend    #    7.83% backend cycles idle      ( +-  0.79% )  (83.33%)
    75,914,422,057      instructions              #    1.39  insn per cycle         
                                                  #    0.06  stalled cycles per insn  ( +-  0.01% )  (83.34%)
    17,027,001,987      branches                  # 1079562962.331 M/sec              ( +-  0.00% )  (83.34%)
       149,974,729      branch-misses             #    0.88% of all branches          ( +-  0.99% )  (83.33%)

           15.7993 +- 0.0205 seconds time elapsed  ( +-  0.13% )
```

After:

```
 Performance counter stats for 'opensim-cmd run-tool setup_forward_tool.xml' (8 runs):

         14,599.66 msec task-clock                #    0.998 CPUs utilized            ( +-  0.28% )
             1,725      context-switches          #  118.131 M/sec                    ( +-  7.63% )
                 0      cpu-migrations            #    0.034 M/sec                    ( +- 53.45% )
             7,541      page-faults               #  516.559 M/sec                    ( +-  0.02% )
    50,510,175,204      cycles                    # 3459778.770 GHz                   ( +-  0.28% )  (83.33%)
     3,956,813,658      stalled-cycles-frontend   #    7.83% frontend cycles idle     ( +-  1.70% )  (83.34%)
     3,275,969,024      stalled-cycles-backend    #    6.49% backend cycles idle      ( +-  0.83% )  (83.32%)
    69,061,193,329      instructions              #    1.37  insn per cycle         
                                                  #    0.06  stalled cycles per insn  ( +-  0.01% )  (83.34%)
    15,753,604,508      branches                  # 1079069439.064 M/sec              ( +-  0.01% )  (83.35%)
       128,656,174      branch-misses             #    0.82% of all branches          ( +-  0.95% )  (83.33%)

           14.6337 +- 0.0385 seconds time elapsed  ( +-  0.26% )
```

Which is, approximately, a ~7 % speedup for that simulation, which is exactly within expectations (I mis-remembered the model that had a 7 % hit: it was `ToyDropLanding`, not `passive_dynamic`).