@adamkewley 2020-06-11
repeats = 16
|                  Test Name | master [secs] | σ [secs] | branch [secs] | σ [secs] | Speedup |
| -------------------------- | ------------- | -------- | ------------- | -------- | ------- |
|                   Gait2354 |          0.26 |     0.00 |          0.26 |     0.00 |    1.00 |
|             ToyDropLanding |         15.14 |     0.04 |         14.26 |     0.10 |    1.06 |
|   ToyDropLanding_nomuscles |          0.44 |     0.00 |          0.44 |     0.00 |    1.00 |
|            passive_dynamic |          4.95 |     0.02 |          5.03 |     0.03 |    0.98 |
| passive_dynamic_noanalysis |          3.15 |     0.01 |          3.20 |     0.01 |    0.98 |
|                      Arm26 |          0.30 |     0.00 |          0.29 |     0.00 |    1.06 |
|             RajagopalModel |         15.28 |     0.05 |         15.13 |     0.13 |    1.01 |

Finally got around to making my scripts dump out tables :D